rename service/ to services/
