+++ src/app.ts
@@ -6,17 +6,16 @@ import { Container } from 'typedi';
 
 import * as env from '@/environment';
 
-import DB from './service/DB';
+import DB from './services/DB';
 import KoaServer from './server/KoaServer';
 import ApolloServerKoa from './server/ApolloServerKoa';
-import rootLogger from './service/logger/rootLogger';
+import rootLogger from './services/logger/rootLogger';
 
 const db = Container.get<DB>(DB);
 const koaServer = Container.get<KoaServer>(KoaServer);
 const apolloServerKoa = Container.get<ApolloServerKoa>(ApolloServerKoa);
 
 (async () => {
-
   await db.connect();
 
   const httpServer = await koaServer.initializeServer();
@@ -30,7 +29,6 @@ const apolloServerKoa = Container.get<ApolloServerKoa>(ApolloServerKoa);
   httpServer.listen(port, () => {
     rootLogger.info(`🚀 HTTP Server ready at port ${port}`);
   });
-
 })().catch(error => {
   rootLogger.error(error);
   process.exit(1);
