+++ src/services/logger/Logger.ts
@@ -1,4 +1,3 @@
-
 import { DEVELOPMENT } from '@/environment';
 
 export enum LogLevel {
