+++ src/services/logger/senders/StdoutSender.ts
